backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic==2.10.3
jinja2>=3.1.2
openai>=1.0.0
python-dotenv>=1.0.0
instructor>=1.0.0
supabase>=2.0.0
pytest>=7.0.0
httpx>=0.24.0
qdrant-client>=1.7.0
python-multipart>=0.0.6
rank-bm25>=0.2.2
cachetools>=5.3.0
ruff>=0.1.0

# LlamaIndex Core & Integrations
llama-index>=0.11.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0
llama-index-vector-stores-qdrant>=0.3.0
llama-index-readers-file>=0.2.0
